feat: pelotas exclusive route
