+++ .nvmrc
@@ -1,0 +1,1 @@
+v20.11.1
