+++ pages/pelotas.vue
@@ -1,0 +1,7 @@
+<template>
+  <MainPage :map-center="[-52.336460593718044, -31.769529722766524]" :map-zoom="11.5" />
+</template>
+
+<style lang="scss">
+//
+</style>
